// client/src/components/pages/titleSearch/types.ts
export type TmdbMovieTitleResponse = {
  page: number;
  results: Array<TmdbMovieRecord>;
  total_pages: number;
  total_results: number;
};

export type TmdbMovieRecord = {
  adult: boolean;
  backdrop_path: string;
  genre_ids: Array<number>;
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};